Complete Q value iteration
The iteration only took into account the direct predecessors of a reward
in a given chain of actions, but did not update states that were not
touched during the round but the final state could be reached from them
with a (series of) valid actions.

The value update now looks for every possible predecessor and initializes
them if missing, and updates values.